Added back old tests for preserve.

+++ tools/clang/test/HLSLFileCheck/dxil/debug/new_noops_no_fold.hlsl
@@ -1,4 +1,4 @@
-// RUN: %dxc -preserve-intermediate-values -E main -T ps_6_0 %s -Od | FileCheck %s
+// RUN: %dxc -E main -T ps_6_0 %s -Od | FileCheck %s
 
 // Test that non-const arithmetic are not optimized away
 
@@ -7,24 +7,28 @@ Texture2D tex1 : register(t1);
 
 [RootSignature("DescriptorTable(SRV(t0), SRV(t1))")]
 float4 main() : SV_Target {
-  // CHECK: %[[p_load:[0-9]+]] = load i32, i32*
-  // CHECK-SAME: @dx.preserve.value
-  // CHECK: %[[p:[0-9]+]] = trunc i32 %[[p_load]] to i1
+  // xHECK: %[[p_load:[0-9]+]] = load i32, i32*
+  // xHECK-SAME: @dx.preserve.value
+  // xHECK: %[[p:[0-9]+]] = trunc i32 %[[p_load]] to i1
 
-  double x = 10;
-  // select i1 %[[p]], double 1.000000e+01, %[[preserve_f64]]
+  float x = 10;
+  // select i1 %[[p]], float 1.000000e+01, float 1.000000e+01
+  // CHECK: dx.nothing
 
-  double y = x + 5;
+  float y = x + 5;
   // CHECK: %[[a1:.+]] = fadd
-  // select i1 %[[p]], double [[a1]], double [[a1]]
+  // select i1 %[[p]], float [[a1]], float [[a1]]
+  // CHECK: dx.nothing
 
-  double z = y * 2;
+  float z = y * 2;
   // CHECK: %[[b1:.+]] = fmul
-  // select i1 %[[p]], double [[b1]], double [[b1]]
+  // select i1 %[[p]], float [[b1]], float [[b1]]
+  // CHECK: dx.nothing
 
-  double w = z / 0.5;
+  float w = z / 0.5;
   // CHECK: %[[c1:.+]] = fdiv
-  // select i1 %[[p]], double [[c1]], double [[c1]]
+  // select i1 %[[p]], float [[c1]], float [[c1]]
+  // CHECK: dx.nothing
 
   Texture2D tex = tex0; 
   // CHECK: load i32, i32*
